API Reference

SPHKernels.Cubic

SPHKernels.Quintic

SPHKernels.SPHKernel

SPHKernels.WendlandC2

SPHKernels.WendlandC4

SPHKernels.WendlandC6

SPHKernels.WendlandC8

SPHKernels.bias_correction_1D

SPHKernels.bias_correction_1D

SPHKernels.bias_correction_1D

SPHKernels.bias_correction_1D

SPHKernels.bias_correction_1D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.kernel_deriv_1D

SPHKernels.kernel_deriv_1D

SPHKernels.kernel_deriv_1D

SPHKernels.kernel_deriv_1D

SPHKernels.kernel_deriv_1D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_value_1D

SPHKernels.kernel_value_1D

SPHKernels.kernel_value_1D

SPHKernels.kernel_value_1D

SPHKernels.kernel_value_1D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.δρ₁

SPHKernels.δρ₂

SPHKernels.δρ₃

SPHKernels.∇𝒲₁

SPHKernels.∇𝒲₂

SPHKernels.∇𝒲₃

SPHKernels.𝒲₁

SPHKernels.𝒲₂

SPHKernels.𝒲₃

Exported Types

SPHKernels.Cubic — Type

struct Cubic <: SPHKernel
    n_neighbours::Int64
    norm_1D::Float64
    norm_2D::Float64
    norm_3D::Float64
    function Cubic(n_neighbours::Int64=64)
        new(n_neighbours, 8.0/π, 8.0/π)
    end
end

Datatype for cubic sph spline.

SPHKernels.Quintic — Type

struct Quintic <: SPHKernel
    n_neighbours::Int64
    norm_1D::Float64
    norm_2D::Float64
    norm_3D::Float64
end

Datatype for quintic sph spline.

SPHKernels.SPHKernel — Type

SPHKernel

Supertype for all SPH kernels.

SPHKernels.WendlandC2 — Type

struct WendlandC2 <: SPHKernel
    n_neighbours::Int64
    norm_1D::Float64
    norm_2D::Float64
    norm_3D::Float64
end

SPHKernels.WendlandC4 — Type

struct WendlandC4 <: SPHKernel
    n_neighbours::Int64
    norm_1D::Float64
    norm_2D::Float64
    norm_3D::Float64
end

SPHKernels.WendlandC6 — Type

struct WendlandC6 <: SPHKernel
    n_neighbours::Int64
    norm_1D::Float64
    norm_2D::Float64
    norm_3D::Float64
end

SPHKernels.WendlandC8 — Type

struct WendlandC8 <: SPHKernel
    n_neighbours::Int64
    norm_2D::Float64
    norm_3D::Float64
    function WendlandC6(n_neighbours::Integer=395)
        new(n_neighbours, 8.0/(3π), 357.0/(64π))
    end
end

Exported Functions

SPHKernels.bias_correction_1D — Method

bias_correction_1D(kernel::Cubic, density::Real, m::Real, h_inv::Real)

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_1D — Method

bias_correction_1D(kernel::Quintic, density::Real, m::Real, h_inv::Real)

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_1D — Method

bias_correction_1D(kernel::WendlandC2, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_1D — Method

bias_correction_1D(kernel::WendlandC4, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_1D — Method

bias_correction_1D(kernel::WendlandC6, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::Cubic, density::Real, m::Real, h_inv::Real)

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::Quintic, density::Real, m::Real, h_inv::Real)

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::WendlandC2, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::WendlandC4, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::WendlandC6, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::WendlandC8, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::Cubic, density::Real, m::Real, h_inv::Real)

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::Quintic, density::Real, m::Real, h_inv::Real)

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::WendlandC2, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::WendlandC4, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::WendlandC6, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::WendlandC8, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.kernel_deriv_1D — Method

kernel_deriv_1D(kernel::Cubic, u::Real, h_inv::Real)

Evaluate the derivative of the Cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_1D — Method

kernel_deriv_1D(kernel::Quintic, u::Real, h_inv::Real)

Evaluate the derivative of the Quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_1D — Method

kernel_deriv_2D(kernel::WendlandC2, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_1D — Method

kernel_deriv_2D(kernel::WendlandC4, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_1D — Method

kernel_deriv_1D(kernel::WendlandC6, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::Cubic, u::Real, h_inv::Real)

Evaluate the derivative of the Cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::Quintic, u::Real, h_inv::Real)

Evaluate the derivative of the Quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::WendlandC2, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::WendlandC4, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::WendlandC6, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::WendlandC8, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::Cubic, u::Real, h_inv::Real)

Evaluate the derivative of the Cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::Quintic, u::Real, h_inv::Real)

Evaluate the derivative of the Quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::WendlandC2, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::WendlandC4, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::WendlandC6, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::WendlandC8, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC8 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_1D — Method

kernel_value_1D(kernel::Cubic, u::Real, h_inv::Real)

Evaluate cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_1D — Method

kernel_value_1D(kernel::Quintic, u::Real, h_inv::Real)

Evaluate quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_1D — Method

kernel_value_1D(kernel::WendlandC2, u::Real, h_inv::Real)

Evaluate WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_1D — Method

kernel_value_1D(kernel::WendlandC4, u::Real, h_inv::Real)

Evaluate WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_1D — Method

kernel_value_1D(kernel::WendlandC6, u::Real, h_inv::Real)

Evaluate WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::Cubic, u::Real, h_inv::Real)

Evaluate cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::Quintic, u::Real, h_inv::Real)

Evaluate quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::WendlandC2, u::Real, h_inv::Real)

Evaluate WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::WendlandC4, u::Real, h_inv::Real)

Evaluate WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::WendlandC6, u::Real, h_inv::Real)

Evaluate WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::WendlandC8, u::Real, h_inv::Real)

Evaluate WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::Cubic, u::Real, h_inv::Real)

Evaluate cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::Quintic, u::Real, h_inv::Real)

Evaluate quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC2, u::Real, h_inv::Real)

Evaluate WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC4, u::Real, h_inv::Real)

Evaluate WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC6, u::Real, h_inv::Real)

Evaluate WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC8, u::Real, h_inv::Real)

Evaluate WendlandC8 spline at position $u = \frac{x}{h}$.

SPHKernels.δρ₁ — Method

δρ₁(kernel::SPHKernel, density::Real, m::Real, h_inv::Real)

Corrects the 1D density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.δρ₂ — Method

δρ₂(kernel::SPHKernel, density::Real, m::Real, h_inv::Real)

Corrects the 2D density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.δρ₃ — Method

δρ₃(kernel::SPHKernel, density::Real, m::Real, h_inv::Real)

Corrects the 3D density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.∇𝒲₁ — Method

∇𝒲₁( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 1D derivative at position $u = \frac{x}{h}$.

SPHKernels.∇𝒲₂ — Method

∇𝒲₂( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 1D derivative at position $u = \frac{x}{h}$.

SPHKernels.∇𝒲₃ — Method

∇𝒲₃( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 1D derivative at position $u = \frac{x}{h}$.

SPHKernels.𝒲₁ — Method

𝒲₁( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 1D spline at position $u = \frac{x}{h}$.

SPHKernels.𝒲₂ — Method

𝒲₂( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 2D spline at position $u = \frac{x}{h}$.

SPHKernels.𝒲₃ — Method

𝒲₃( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 3D spline at position $u = \frac{x}{h}$.

Private Functions

Private Types