API Reference

SPHKernels.Cubic

SPHKernels.Quintic

SPHKernels.SPHKernel

SPHKernels.WendlandC4

SPHKernels.WendlandC6

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

Exported Types

SPHKernels.Cubic — Type

struct Cubic <: SPHKernel
    n_neighbours::Int64
    norm_2D::Float64
    norm_3D::Float64
    function Cubic(n_neighbours::Int64=64)
        new(n_neighbours, 8.0/π, 8.0/π)
    end
end

Datatype for cubic sph spline.

SPHKernels.Quintic — Type

struct Quintic <: SPHKernel
    n_neighbours::Int64
    norm_2D::Float64
    norm_3D::Float64
    function Quintic(n_neighbours::Integer=216)
        new(n_neighbours, 15309.0/(478.0*π), 2187.0/(40.0*pi))
    end
end

Datatype for quintic sph spline.

SPHKernels.SPHKernel — Type

SPHKernel

Supertype for all SPH kernels.

SPHKernels.WendlandC4 — Type

struct WendlandC4 <: SPHKernel
    n_neighbours::Int64
    norm_2D::Float64
    norm_3D::Float64
    function WendlandC4(n_neighbours::Int64=216)
        new(n_neighbours,9.0/π, 495.0/(32.0 * π))
    end
end

SPHKernels.WendlandC6 — Type

struct WendlandC6 <: SPHKernel
    n_neighbours::Int64
    norm_2D::Float64
    norm_3D::Float64
    function WendlandC6(n_neighbours::Int64=295)
        new(n_neighbours, 78.0/(7.0*π), 1365.0/(64.0*π))
    end
end

Exported Functions

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::Cubic, density::Real, m::Real, h_inv::Real)

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::Quintic, density::Real, m::Real, h_inv::Real)

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::WendlandC4, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::WendlandC6, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::Cubic, density::Real, m::Real, h_inv::Real)

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::Quintic, density::Real, m::Real, h_inv::Real)

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::WendlandC4, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::WendlandC6, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::Cubic, u::Real, h_inv::Real)

Evaluate the derivative of the Cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::Quintic, u::Real, h_inv::Real)

Evaluate the derivative of the Quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::WendlandC4, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::WendlandC6, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::Cubic, u::Real, h_inv::Real)

Evaluate the derivative of the Cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::Quintic, u::Real, h_inv::Real)

Evaluate the derivative of the Quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::WendlandC4, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::WendlandC6, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::Cubic, u::Real, h_inv::Real)

Evaluate cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::Quintic, u::Real, h_inv::Real)

Evaluate quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::WendlandC4, u::Real, h_inv::Real)

Evaluate WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::WendlandC6, u::Real, h_inv::Real)

Evaluate WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::Cubic, u::Real, h_inv::Real)

Evaluate cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::Quintic, u::Real, h_inv::Real)

Evaluate quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC4, u::Real, h_inv::Real)

Evaluate WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC6, u::Real, h_inv::Real)

Evaluate WendlandC6 spline at position $u = \frac{x}{h}$.

Private Functions

Private Types