API Reference

SPHKernels.Cubic

SPHKernels.Quintic

SPHKernels.SPHKernel

SPHKernels.WendlandC2

SPHKernels.WendlandC2

SPHKernels.WendlandC4

SPHKernels.WendlandC6

SPHKernels.WendlandC8

SPHKernels.bias_correction_1D

SPHKernels.bias_correction_1D

SPHKernels.bias_correction_1D

SPHKernels.bias_correction_1D

SPHKernels.bias_correction_1D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_2D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.bias_correction_3D

SPHKernels.kernel_deriv_1D

SPHKernels.kernel_deriv_1D

SPHKernels.kernel_deriv_1D

SPHKernels.kernel_deriv_1D

SPHKernels.kernel_deriv_1D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_2D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_deriv_3D

SPHKernels.kernel_value_1D

SPHKernels.kernel_value_1D

SPHKernels.kernel_value_1D

SPHKernels.kernel_value_1D

SPHKernels.kernel_value_1D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_2D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.kernel_value_3D

SPHKernels.δρ₁

SPHKernels.δρ₂

SPHKernels.δρ₃

SPHKernels.∇𝒲₁

SPHKernels.∇𝒲₂

SPHKernels.∇𝒲₃

SPHKernels.𝒲₁

SPHKernels.𝒲₂

SPHKernels.𝒲₃

Exported Types

SPHKernels.Cubic — Type

Cubic(T::DataType=Float64, n_neighbours::Integer=64)

Set up a Cubic kernel for a given DataType T.

SPHKernels.Quintic — Type

Quintic(T::DataType=Float64, n_neighbours::Integer=64)

Set up a Quintic kernel for a given DataType T.

SPHKernels.SPHKernel — Type

SPHKernel

Supertype for all SPH kernels.

SPHKernels.WendlandC2 — Type

struct WendlandC2{T} <: SPHKernel
    n_neighbours::Int64
    norm_1D::T
    norm_2D::T
    norm_3D::T
end

SPHKernels.WendlandC2 — Type

WendlandC2(T::DataType=Float64, n_neighbours::Integer=100)

Set up a WendlandC2 with a given DataType T.

SPHKernels.WendlandC4 — Type

WendlandC4(T::DataType=Float64, n_neighbours::Integer=216)

Set up a WendlandC4 kernel for a given DataType T.

SPHKernels.WendlandC6 — Type

WendlandC6(T::DataType=Float64, n_neighbours::Integer=295)

Set up a WendlandC6 kernel for a given DataType T.

SPHKernels.WendlandC8 — Type

WendlandC8(T::DataType=Float64, n_neighbours::Integer=395)

Set up a WendlandC8 kernel for a given DataType T.

Exported Functions

SPHKernels.bias_correction_1D — Method

bias_correction_1D(kernel::Cubic{T}, density::Real, m::Real, h_inv::Real) where T

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_1D — Method

bias_correction_1D(kernel::Quintic{T}, density::Real, m::Real, h_inv::Real) where T

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_1D — Method

bias_correction_1D(kernel::WendlandC2{T}, density::Real, m::Real, h_inv::Real) where T

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_1D — Method

bias_correction_1D(kernel::WendlandC4{T}, density::Real, m::Real, h_inv::Real) where T

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_1D — Method

bias_correction_1D(kernel::WendlandC6, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::Cubic, density::Real, m::Real, h_inv::Real)

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::Quintic{T}, density::Real, m::Real, h_inv::Real) where T

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::WendlandC2{T}, density::Real, m::Real, h_inv::Real) where T

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::WendlandC4, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::WendlandC6{T}, density::Real, m::Real, h_inv::Real) where T

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_2D — Method

bias_correction_2D(kernel::WendlandC8, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::Cubic{T}, density::Real, m::Real, h_inv::Real) where T

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::Quintic{T}, density::Real, m::Real, h_inv::Real) where T

Does not do anything for the BSplines. Implemented for stability.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::WendlandC2, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::WendlandC4, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::WendlandC6, density::Real, m::Real, h_inv::Real)

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.bias_correction_3D — Method

bias_correction_3D(kernel::WendlandC8{T}, density::Real, m::Real, h_inv::Real) where T

Corrects the density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.kernel_deriv_1D — Method

kernel_deriv_1D(kernel::Cubic{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the Cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_1D — Method

kernel_deriv_1D(kernel::Quintic{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the Quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_1D — Method

kernel_deriv_1D(kernel::WendlandC2{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_1D — Method

kernel_deriv_1D(kernel::WendlandC4{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_1D — Method

kernel_deriv_1D(kernel::WendlandC6{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::Cubic{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the Cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::Quintic{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the Quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::WendlandC2{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::WendlandC4{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::WendlandC6, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_2D — Method

kernel_deriv_2D(kernel::WendlandC8{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::Cubic{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the Cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::Quintic, u::Real, h_inv::Real)

Evaluate the derivative of the Quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::WendlandC2{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::WendlandC4{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::WendlandC6{T}, u::Real, h_inv::Real) where T

Evaluate the derivative of the WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_deriv_3D — Method

kernel_deriv_3D(kernel::WendlandC8, u::Real, h_inv::Real)

Evaluate the derivative of the WendlandC8 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_1D — Method

kernel_value_1D(kernel::Cubic{T}, u::Real, h_inv::Real) where T

Evaluate cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_1D — Method

kernel_value_1D(kernel::Quintic{T}, u::Real, h_inv::Real) where T

Evaluate quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_1D — Method

kernel_value_1D(kernel::WendlandC2{T}, u::Real, h_inv::Real) where T

Evaluate WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_1D — Method

kernel_value_1D(kernel::WendlandC4{T}, u::Real, h_inv::Real) where T

Evaluate WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_1D — Method

kernel_value_1D(kernel::WendlandC6{T}, u::Real, h_inv::Real) where T

Evaluate WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::Cubic{T}, u::Real, h_inv::Real) where T

Evaluate cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::Quintic{T}, u::Real, h_inv::Real) where T

Evaluate quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::WendlandC2{T}, u::Real, h_inv::Real) where T

Evaluate WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

(kernel::WendlandC4{T}, u::Real, h_inv::Real) where T

Evaluate WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

(kernel::WendlandC6{T}, u::Real, h_inv::Real) where T

Evaluate WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_2D — Method

kernel_value_2D(kernel::WendlandC8{T}, u::Real, h_inv::Real) where T

Evaluate WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::Cubic{T}, u::Real, h_inv::Real) where T

Evaluate cubic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::Quintic{T}, u::Real, h_inv::Real) where T

Evaluate quintic spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC2{T}, u::Real, h_inv::Real) where T

Evaluate WendlandC2 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC4{T}, u::Real, h_inv::Real) where T

Evaluate WendlandC4 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC6{T}, u::Real, h_inv::Real) where T

Evaluate WendlandC6 spline at position $u = \frac{x}{h}$.

SPHKernels.kernel_value_3D — Method

kernel_value_3D(kernel::WendlandC8{T}, u::Real, h_inv::Real) where T

Evaluate WendlandC8 spline at position $u = \frac{x}{h}$.

SPHKernels.δρ₁ — Method

δρ₁(kernel::SPHKernel, density::Real, m::Real, h_inv::Real)

Corrects the 1D density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.δρ₂ — Method

δρ₂(kernel::SPHKernel, density::Real, m::Real, h_inv::Real)

Corrects the 2D density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.δρ₃ — Method

δρ₃(kernel::SPHKernel, density::Real, m::Real, h_inv::Real)

Corrects the 3D density estimate for the kernel bias. See Dehnen&Aly 2012, eq. 18+19.

SPHKernels.∇𝒲₁ — Method

∇𝒲₁( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 1D derivative at position $u = \frac{x}{h}$.

SPHKernels.∇𝒲₂ — Method

∇𝒲₂( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 1D derivative at position $u = \frac{x}{h}$.

SPHKernels.∇𝒲₃ — Method

∇𝒲₃( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 1D derivative at position $u = \frac{x}{h}$.

SPHKernels.𝒲₁ — Method

𝒲₁( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 1D spline at position $u = \frac{x}{h}$.

SPHKernels.𝒲₂ — Method

𝒲₂( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 2D spline at position $u = \frac{x}{h}$.

SPHKernels.𝒲₃ — Method

𝒲₃( kernel::SPHKernel, u::Real, h_inv::Real)

Evaluate 3D spline at position $u = \frac{x}{h}$.

Private Functions

Private Types